+++ examples/03-arguments.lua
@@ -1,0 +1,4 @@
+-- You can pass arguments to the script by passing extra values to remodel!
+-- Try running this example as `remodel examples/03-arguments.lua hello`
+
+print("My arguments:", ...)