+++ src/main.rs
@@ -1,6 +1,6 @@
 use std::{error::Error, fs, path::PathBuf};
 
-use rlua::Lua;
+use rlua::{Lua, MultiValue, ToLua};
 use structopt::StructOpt;
 
 mod lua_api;
@@ -11,6 +11,8 @@ use lua_api::Remodel;
 struct Options {
     #[structopt(parse(from_os_str))]
     script: PathBuf,
+
+    script_arguments: Vec<String>,
 }
 
 fn start() -> Result<(), Box<dyn Error>> {
@@ -19,10 +21,16 @@ fn start() -> Result<(), Box<dyn Error>> {
     let lua = Lua::new();
 
     lua.context(move |context| {
+        let lua_args = opt
+            .script_arguments
+            .into_iter()
+            .map(|value| value.to_lua(context))
+            .collect::<Result<Vec<_>, _>>()?;
+
         context.globals().set("remodel", Remodel)?;
 
         let chunk = context.load(&contents);
-        chunk.exec()
+        chunk.call(MultiValue::from_vec(lua_args))
     })?;
 
     Ok(())